Revert "Feature/8 handle routes for signup"
