+++ talenthive-website/backend/.env
@@ -1,0 +1,1 @@
+PORT = 4000; // Port number for the server
