--- talenthive-website/backend/.env.example
@@ -1,6 +1,0 @@
-PORT=
-HOST=
-
-MONGO_USER=
-MONGO_PASSWORD=
-MONGO_URI=
