+++ talenthive-website/backend/controllers/user.ts
@@ -1,0 +1,2 @@
+export const createUser = async (req: Request, res: Response) => {
+};
