+++ talenthive-website/backend/database/index.ts
@@ -1,0 +1,3 @@
+import mongoose from 'mongoose';
+
+// Connect to MongoDB
