+++ talenthive-website/backend/models/user.ts
@@ -1,0 +1,20 @@
+import mongoose from "mongoose";
+
+// Example
+const User = new mongoose.Schema({
+    email: {
+        type: String,
+        required: true,
+        unique: true,
+    },
+    password: {
+        type: String,
+        required: true,
+    },
+    role: {
+        type: String,
+        required: true,
+    },
+    });
+
+export default mongoose.model('User', User);
