+++ talenthive-website/backend/package.json
@@ -1,29 +1,22 @@
 {
   "name": "talenthive-website",
   "version": "1.0.0",
-  "main": "src/server.ts",
+  "main": "index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "ts-node-dev  ./src/server.ts --respawn --transpileOnly"
+    "start": "node index.ts"
   },
   "author": "",
   "license": "ISC",
   "description": "",
   "devDependencies": {
-    "@types/cors": "^2.8.17",
-    "@types/express": "^5.0.0",
-    "@types/morgan": "^1.9.9",
     "@types/node": "^22.8.1",
-    "ts-node-dev": "^2.0.0",
-    "typescript": "^5.6.3"
+    "nodemon": "^3.1.7"
   },
   "dependencies": {
-    "@types/validator": "^13.12.2",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.21.1",
-    "mongoose": "^8.7.3",
-    "morgan": "^1.10.0",
-    "validator": "^13.12.0"
+    "mongoose": "^8.7.3"
   }
 }
