+++ talenthive-website/backend/routes/index.ts
@@ -1,0 +1,10 @@
+import express from 'express';
+import userRouter from './user';
+
+const router = express.Router();
+
+router.use('/user', userRouter);
+
+
+
+export default router;
