+++ talenthive-website/backend/routes/user.ts
@@ -1,0 +1,10 @@
+import express from 'express';
+
+import { createUser } from '../controllers/user';
+
+const router = express.Router();
+
+router.post('/user', createUser); // example route
+
+
+export default router;
