--- talenthive-website/backend/src/controllers/user.ts
@@ -1,79 +1,0 @@
-import { Request, Response } from "express";
-import User from "../models/user";
-
-const getAllUsers = async (req: Request, res: Response) => {
-    try {
-        const users = await User.find();
-        res.status(200).json({
-            status: 'success',
-            data: users
-        });
-    } catch (err) {
-        res.status(500).json({
-            status: 'error',
-            message: err
-        });
-    }
-}
-
-const getUserById = async (req: Request, res: Response) => {
-    try {
-        const user = await User.findById(req.params.id);
-        res.status(200).json({
-            status: 'success',
-            data: user
-        });
-    } catch (err) {
-        res.status(500).json({
-            status: 'error',
-            message: err
-        })
-    }
-}
-
-const createUser = async (req: Request, res: Response) => {
-    try {
-        const user = await User.create(req.body);
-        res.status(201).json({
-            status: 'success',
-            data: user
-        });
-    } catch (err) {
-        res.status(500).json({
-            status: 'error',
-            message: err
-        });
-    }
-}
-
-const updateUser = async (req: Request, res: Response) => {
-    try {
-        const user = await User.findByIdAndUpdate(req.params);
-        res.status(200).json({
-            status: 'success',
-            data: user
-        });
-    } catch (err) {
-        res.status(500).json({
-            status: 'error',
-            message: err
-        });
-    }
-}
-
-const deleteUser = async (req: Request, res: Response) => {
-    try {
-        await User.findByIdAndDelete(req.params.id);
-        res.status(204).json({
-            status: 'success',
-            data: null
-        });
-    } catch (err) {
-        res.status(500).json({
-            status: 'error',
-            message: err
-        });
-    }
-}
-
-export { getAllUsers, getUserById, createUser, updateUser, deleteUser };
