--- talenthive-website/backend/src/database/config.ts
@@ -1,7 +1,0 @@
-import mongoose from "mongoose";
-
-const DB = process.env
-    .MONGO_URI!.replace("<MONGO_PASSWORD>", process.env.MONGO_PASSWORD!)
-    .replace("<MONGO_USER>", process.env.MONGO_USER!);
-
-mongoose.connect(DB, {}).then(() => console.log("DB connection successful!"));
