--- talenthive-website/backend/src/models/rolePermisstion.ts
@@ -1,26 +1,0 @@
-import { Schema, Types, model } from "mongoose";
-
-
-interface IRolePermisstion {
-    _id: Types.ObjectId;
-    role_id: Types.ObjectId;
-    permission_id: Types.ObjectId;
-}
-
-const RolePermisstionSchema = new Schema<IRolePermisstion>({
-    role_id: {
-        type: Schema.Types.ObjectId,
-        ref: 'Role',
-        required: true,
-
-    },
-    permission_id: {
-        type: Schema.Types.ObjectId,
-        ref: 'Permission',
-        required: true
-    }
-})
-
-const RolePermisstion = model<IRolePermisstion>('RolePermisstion', RolePermisstionSchema);
-
-export default RolePermisstion;
