--- talenthive-website/backend/src/routes/authRoutes.ts
@@ -1,8 +1,0 @@
-import { Router } from "express";
-import * as authController from "../controllers/authController";
-
-const router = Router();
-
-router.post("/register", authController.register);
-
-export default router;
