--- talenthive-website/backend/src/routes/index.ts
@@ -1,9 +1,0 @@
-import { Router } from "express";
-
-const indexRouter = Router();
-
-indexRouter.get("/", (req, res) => {
-  res.send("Hello world");
-});
-
-export default indexRouter;
