+++ talenthive-website/frontend/src/index.css
@@ -1,5 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Noto+Sans:ital,wght@0,100..900;1,100..900&family=Roboto:wght@100;300;400;500;700&display=swap');
-
 *,
 *::before,
 *::after {
@@ -9,51 +7,9 @@
 }
 
 :root {
-    --primary-color: #3493FF;
-    --text-color: #000;
-    --button-color: #fff;
-    --form-color: #fff;
+    --primay-color: #3493FF;
     --border-color: #d9d9d9;
-    --box-shadow-item: 4px 4px 4px rgb(0, 0, 0, 25%);
-    --box-shadow-button: 0 4px 4px rgb(0, 0, 0, 25%);
+    --box-shadow-color: 4px 4px 4px rgb(0 0 0 25%);
     --background-gradient: linear-gradient(to right, #1A4A81, var(--primay-color));
-    --body-gradient: linear-gradient(to bottom, #F2FCFF, #FAF6FF);
-
-
-    --margin-element-form: 20px;
-    --border-radius-inp: 4px;
-    --border-radius-btn: 8px;
-    --border-radius-form: 10px;
 }
-
-html {
-  
-}
-
-body {
-  font-family: 'Noto Sans', 'Roboto', sans-serif;
-  min-height: 1000px;
-  background-image: var(--body-gradient);
-  background-repeat: no-repeat;
-}
-
-button:not(:disabled) {
-  cursor: pointer;
-  transition: opacity 0.15s;
-
-  &:hover {
-    opacity: 0.7;
-  }
-
-  &:active {
-    opacity: 0.3;
-  }
-}
-
-button:disabled {
-  background-color: var(--border-color);
-  cursor: default;
-}
-
-
