--- talenthive-website/frontend/src/styles/layouts/AuthenticationLayout.module.css
@@ -1,9 +1,0 @@
-.wrapper {
-    display: flex;
-    flex-direction: column;
-    margin: 100px 200px;
-}
-
-.form {
-    margin: 5% 10%;
-}
