--- talenthive-website/frontend/src/styles/pages/Authentication.module.css
@@ -1,107 +1,0 @@
-.wrapper {
-    display: flex;
-}
-
-.form {
-    background-color: var(--form-color);
-    flex: 1;
-    border: 1px solid var(--primary-color);
-    border-radius: var(--border-radius-form);
-    height: auto;
-    min-width: 700px;
-    padding: 40px 100px;
-    box-shadow: var(--box-shadow-item);
-    display: flex;
-    flex-direction: column;
-
-
-    input {
-        border: 1px solid var(--border-color);
-        border-radius: var(--border-radius-inp);
-        margin-top: 5px;
-        padding: 10px;
-        width: 100%;
-        height: 40px;
-
-        &.checkbox-input {
-            width: 20px;
-            border: 1px solid var(--border-color);
-            margin: 0px;
-            margin-right: 10px;
-        }
-    }
-    
-
-    span {
-        color: red;
-    }
-
-
-    &>*{
-        margin-bottom: var(--margin-element-form);
-    }
-}
-
-.link {
-    text-decoration: none;
-    color: var(--primary-color);
-}
-
-.email, .password {
-    display: flex;
-    flex-direction: column;
-}
-
-li {
-        
-    font-size: .8rem;
-    margin-left: 20px;
-
-    &.none{
-        color: var(--text-color);
-    }
-
-    &.valid {
-        color: #0ab305;
-    }
-
-    &.invalid {
-        color: #f60d00;
-    }
-}
-
-.policy-content {
-    flex: 1;
-}
-
-.btn {
-    padding: 15px 25px;
-    width: auto;
-    font-weight: 700;
-    font-size: 18px;
-    background-color: var(--primary-color);
-    color: var(--button-color);
-    border: none;
-    border-radius: var(--border-radius-btn);
-    box-shadow: var(--box-shadow-button);
-}
-
-.question {
-    margin-right: 5px;
-}
-
-.signin, .signup{
-    margin-bottom: 0px;
-}
-
-.signup-link, .signin-link {
-    font-weight: 600;
-}
-
-.container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-
